move grooveshark API to api for cleanup.
